Update conversion_factor_enum.py
I added the conversion constant and conversion offset for the analog accelerometer to the conversion_factor_enum. These represent the accelerometer's sensitivity factor and zero G voltage, respectively, and are pulled from page 3 of the manufacturer's datasheet. The sensitivity factor is that a change of 0.3 volts represents a change of 1 g of acceleration. The zero G voltage is that when the accelerometer is not experiencing any acceleration, it will read 1.5 volts in all three axes. 

I also added a conversion factor for converting between Gs and m/s^2. One G = 9.81 m/s^2